+++ src/config/system/rebase.ts
@@ -1,6 +1,6 @@
 import UFragments from './abis/UFragments.json';
 import UFragmentsPolicy from './abis/UFragmentsPolicy.json';
-import ChainlinkOracle from './abis/ChainlinkOracle.json';
+import ChainlinkPricefeed from './abis/ChainlinkPricefeed.json';
 import Orchestrator from './abis/Orchestrator.json';
 import { digg_system } from '../deployments/mainnet.json';
 import { RebaseNetworkConfig } from '../../mobx/model';
@@ -30,13 +30,17 @@ export const getRebase = (network?: string | null): RebaseNetworkConfig | undefi
 						namespace: 'policy',
 					},
 					{
-						addresses: [digg_system.chainlinkOracle],
-						abi: ChainlinkOracle.abi as AbiItem[],
+						addresses: [digg_system.chainlinkPriceFeed],
+						abi: ChainlinkPricefeed.abi as AbiItem[],
 						groupByNamespace: true,
 						namespace: 'oracle',
 						readMethods: [
 							{
 								name: 'latestAnswer',
+								args: [],
+							},
+							{
+								name: 'latestTimestamp',
 								args: [],
 							},
 						],