+++ changelog.md
@@ -1,5 +1,10 @@
 # v2 UI Changelog
 
+### Hotfix - 07/05/2021
+
+-   updated digg oracle
+-   fixed an issue causing account balance updates to be delayed 
+
 ### v2.8.1 - 07/01/2021
 
 -   introduced guarded vaults to the app